requirements.txt
langchain==0.2.0
langchain-community==0.2.0
langchain-openai==0.1.7
tavily-python==0.3.3
faiss-cpu==1.8.0
requests>=2.32.2
